+++ frontend/src/pages/Player/Toolbar/DevToolsWindowV2/DevToolsWindowV2.tsx
@@ -8,7 +8,6 @@ import {
 	IconSolidSearch,
 	IconSolidSwitchHorizontal,
 	MenuButton,
-	MultiSelectButton,
 	Tabs,
 	Text,
 	useFormState,
@@ -31,18 +30,16 @@ import {
 	RequestType,
 	Tab,
 } from '@pages/Player/Toolbar/DevToolsWindowV2/utils'
-import {
-	ICountPerRequestStatus,
-	ICountPerRequestType,
-} from '@pages/Player/Toolbar/DevToolsWindowV2/utils'
 import useLocalStorage from '@rehooks/local-storage'
 import clsx from 'clsx'
-import React, { useMemo } from 'react'
+import React from 'react'
 import { Link } from 'react-router-dom'
 
 import { useLinkLogCursor } from '@/pages/Player/PlayerHook/utils'
 import { styledVerticalScrollbar } from '@/style/common.css'
 
+import { RequestTypeFilter } from './RequestTypeFilter/RequestTypeFilter'
+import { RequestStatusFilter } from './RequestStatusFilter/RequestStatusFilter'
 import { ConsolePage } from './ConsolePage/ConsolePage'
 import ErrorsPage from './ErrorsPage/ErrorsPage'
 import * as styles from './style.css'
@@ -287,7 +284,7 @@ const DevToolsWindowV2: React.FC<
 										</>
 									) : selectedDevToolsTab === Tab.Network ? (
 										<>
-											<RequestTypeMultiSelect
+											<RequestTypeFilter
 												requestTypes={requestTypes}
 												setRequestTypes={
 													setRequestTypes
@@ -296,7 +293,7 @@ const DevToolsWindowV2: React.FC<
 													parsedResources
 												}
 											/>
-											<RequestStatusMultiSelect
+											<RequestStatusFilter
 												requestStatuses={
 													requestStatuses
 												}
@@ -369,192 +366,6 @@ const DevToolsWindowV2: React.FC<
 				</Box>
 			)}
 		</ResizePanel>
-	)
-}
-
-const RequestTypeMultiSelect = ({
-	requestTypes,
-	setRequestTypes,
-	parsedResources,
-}: any) => {
-	const handleRequestTypeChange = (valueNames: string[]) => {
-		const allPreviouslySelected = requestTypes.includes(RequestType.All)
-		const allCurrentlySelected = valueNames.includes(RequestType.All)
-		const clearOtherTypes = !allPreviouslySelected && allCurrentlySelected
-		const deselectAllType = allPreviouslySelected && valueNames.length > 1
-
-		if (!valueNames.length || clearOtherTypes) {
-			return setRequestTypes([RequestType.All])
-		}
-
-		//-- Set type to be the requestName value --//
-		const updatedRequestTypes = valueNames.reduce(
-			(agg: string[], name: string) => {
-				if (deselectAllType && name === RequestType.All) {
-					return agg
-				}
-
-				return agg.concat(name)
-			},
-			[] as string[],
-		)
-
-		setRequestTypes(updatedRequestTypes)
-	}
-
-	/* Count request per type (XHR, etc.) */
-	const countPerRequestType = useMemo(() => {
-		const count: ICountPerRequestType = {
-			All: 0,
-			link: 0,
-			script: 0,
-			other: 0,
-			xmlhttprequest: 0,
-			css: 0,
-			iframe: 0,
-			fetch: 0,
-			img: 0,
-		}
-
-		parsedResources.forEach((request: any) => {
-			const requestType =
-				request.initiatorType as keyof ICountPerRequestType
-
-			count['All'] += 1
-			/* Only count request types defined in ICountPerRequestType, e.g. skip 'beacon' */
-			if (count.hasOwnProperty(request.initiatorType)) {
-				count[requestType] += 1
-			}
-		})
-
-		return count
-	}, [parsedResources])
-
-	const options = Object.entries(RequestType).map(
-		([displayName, requestName]) => ({
-			key: requestName,
-			render: `${displayName} (${countPerRequestType[requestName]})`,
-		}),
-	)
-
-	return (
-		<MultiSelectButton
-			label="Type"
-			defaultValue={RequestType.All}
-			value={requestTypes}
-			options={options}
-			onChange={handleRequestTypeChange}
-		/>
-	)
-}
-
-const RequestStatusMultiSelect = ({
-	requestStatuses,
-	setRequestStatuses,
-	parsedResources,
-	requestTypes,
-}: any) => {
-	const handleRequestTypeChange = (valueNames: string[]) => {
-		const allPreviouslySelected = requestStatuses.includes(
-			RequestStatus.All,
-		)
-		const allCurrentlySelected = valueNames.includes(RequestStatus.All)
-		const clearOtherTypes = !allPreviouslySelected && allCurrentlySelected
-		const deselectAllType = allPreviouslySelected && valueNames.length > 1
-
-		if (!valueNames.length || clearOtherTypes) {
-			return setRequestStatuses([RequestStatus.All])
-		}
-
-		//-- Set type to be the requestName value --//		)
-		const updatedRequestStatuses = valueNames.reduce(
-			(agg: string[], name: string) => {
-				if (deselectAllType && name === RequestStatus.All) {
-					return agg
-				}
-
-				return agg.concat(name)
-			},
-			[] as string[],
-		)
-
-		setRequestStatuses(updatedRequestStatuses)
-	}
-
-	/* Count request per type (XHR, etc.) */
-	/* Count request per http status (200, etc.) */
-	const countPerRequestStatus = useMemo(() => {
-		const count: ICountPerRequestStatus = {
-			All: 0,
-			'1XX': 0,
-			'2XX': 0,
-			'3XX': 0,
-			'4XX': 0,
-			'5XX': 0,
-			Unknown: 0,
-		}
-
-		parsedResources
-			.filter(
-				(r: any) =>
-					requestTypes.includes(RequestType.All) ||
-					requestTypes.includes(r.initiatorType),
-			)
-			.forEach((request: any) => {
-				const status: number | undefined =
-					request?.requestResponsePairs?.response?.status
-
-				count['All'] += 1
-				if (status) {
-					switch (true) {
-						case status >= 100 && status < 200:
-							count['1XX'] += 1
-							break
-						case status >= 200 && status < 300:
-							count['2XX'] += 1
-							break
-						case status >= 300 && status < 400:
-							count['3XX'] += 1
-							break
-						case status >= 400 && status < 500:
-							count['4XX'] += 1
-							break
-						case status >= 500 && status < 600:
-							count['5XX'] += 1
-							break
-						default:
-							count['Unknown'] += 1
-							break
-					}
-				} else {
-					// this is a network request with no status code
-					// if fetch, consider unknown. otherwise assume it is 2xx
-					if (request.initiatorType === RequestType.Fetch) {
-						count['Unknown'] += 1
-					} else {
-						count['2XX'] += 1
-					}
-				}
-			})
-
-		return count
-	}, [parsedResources, requestTypes])
-
-	const options = Object.entries(RequestStatus).map(
-		([statusKey, statusValue]) => ({
-			key: statusValue,
-			render: `${statusKey} (${countPerRequestStatus[statusValue]})`,
-		}),
-	)
-
-	return (
-		<MultiSelectButton
-			label="Status"
-			defaultValue={options[0].key}
-			value={requestStatuses}
-			options={options}
-			onChange={handleRequestTypeChange}
-		/>
 	)
 }
